Refactoring tests, fixing bugs in hf compatibility test
